ai_service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
pillow==10.3.0
pytesseract==0.3.10
httpx==0.27.0
